Validate lease compatibility and add metrics
This change adds validation in compatible and upgrade modes to
match stream in the lease record with the stream that the application
is configured with. This validation is done when the shard syncer is
deciding whether a lease exists for a particular shard or not.

This change also adds three new metrics: NumIncompatibleLeases,
NumSingleStreamLeases and NumMultiStreamLeases